requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
joblib==0.16.0
lightgbm==2.3.1
pandas==1.1.0
scikit-learn==0.22.1
scipy==1.5.2
opencv-python==4.1.1.26
numpy==1.17.3
trimesh
pyyaml
SimpleITK
pynrrd